Moved `VectorDefinitions` to assets and `VectorSelector` to framework
Adjusted code using VectorSelector
