+++ frontend/src/framework/components/VectorSelector/index.ts
@@ -1,0 +1,2 @@
+export { VectorSelector, addVectorToVectorSelectorData, createVectorSelectorDataFromVectors } from "./vectorSelector";
+export type { VectorSelectorProps } from "./vectorSelector";
