+++ frontend/src/framework/components/VectorSelector/private-utils/VectorSelection.ts
