--- frontend/src/lib/components/VectorSelector/index.ts
@@ -1,2 +1,0 @@
-export { VectorSelector } from "./vectorSelector";
-export type { VectorSelectorProps } from "./vectorSelector";
